B #5525: Fail on Java OCA build errors (#557)
(cherry picked from commit 801f9cf27f47f25bf4ce91b66ddf5a5a0dbabc8a)

+++ src/oca/java/build.sh
@@ -16,6 +16,8 @@
 # limitations under the License.                                             #
 #--------------------------------------------------------------------------- #
 
+set -e
+
 #-------------------------------------------------------------------------------
 # DIR DEFINITIONS
 #-------------------------------------------------------------------------------
@@ -117,7 +119,10 @@ do_clean()
     mkdir -p $BIN_DIR
 
     find share/examples -name '*.class' -delete
-    find test/ -name '*.class' -delete
+
+    if [ -d test/ ]; then
+        find test/ -name '*.class' -delete
+    fi
 
     echo "Cleaning javadoc files..."
     rm -rf $DOC_DIR > /dev/null 2>&1
